# Documents/api_specification.yaml
openapi: 3.0.3
info:
  title: Speak2MD API
  version: "0.1.0"
  description: |
    API для сервиса автоматической расшифровки аудио и генерации конспектов в Markdown.
    Поддерживает загрузку аудио, отслеживание прогресса, получение результата,
    управление транскриптами и базовую аутентификацию. 
    **Примечания:** 
    - Поддерживаемые форматы входных файлов: mp3, m4a, wav.
    - Базовые лимиты MVP: размер файла до 50 МБ, длительность до 120 минут.
servers:
  - url: https://api.speak2md.example.com
    description: Прод
  - url: http://localhost:8000
    description: Локальная разработка
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UUID:
      type: string
      format: uuid
      example: "9f9a9f8c-4f31-4f7a-9c9b-21b2a80d2b65"
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Недопустимый формат файла"
...
    JobStatusResponse:
      type: object
      required: [status, progress]
      properties:
        status:
          type: string
          description: Текущий статус задачи
          enum: [processing, ready, error]
          example: processing
        progress:
          type: integer
          description: Прогресс обработки в процентах
          minimum: 0
          maximum: 100
          example: 42
...
paths:
...
  /api/status/{job_id}:
    get:
      tags: [Jobs]
      summary: Получить статус обработки задачи
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Текущий статус задачи
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
              examples:
                processing:
                  value: {status: processing, progress: 18}
                ready:
                  value: {status: ready, progress: 100}
                error:
                  value: {status: error, progress: 0}
        "400":
          description: Некорректный идентификатор задачи
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        "401":
          description: Неавторизован
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        "404":
          description: Задача не найдена
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
  /api/result/{job_id}:
    get:
      tags: [Jobs]
      summary: Скачать результат обработки (Markdown или JSON)
      description: |
        По умолчанию возвращает Markdown-файл. Можно указать `format=json` или заголовок `Accept: application/json`.
        Для скачивания как файла можно воспользоваться заголовком `Content-Disposition` на стороне сервера.
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: format
          in: query
          description: Формат ответа
          required: false
          schema:
            type: string
            enum: [markdown, json]
      responses:
        "200":
          description: Успех
          content:
            text/markdown:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: object
                description: JSON-представление результата (если доступно)
                example:
                  title: "Стенограмма совещания"
                  sections:
                    - heading: "Итоги"
                      bullets: ["Релиз сдвинут на 2 недели", "Добавить VAD улучшения"]
        "400":
          description: Некорректные параметры
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        "401":
          description: Неавторизован
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        "404":
          description: Результат не найден
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
        "500":
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Error'}
...
security: []